# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: MasterAnnonce API
  version: "1.0.0"
  description: API REST Jakarta/JAX-RS pour la gestion d'annonces (TP3)
servers:
  - url: http://localhost:8080/tpAIR1/api
paths:
  /params:
    get:
      summary: Exemple QueryParams
      parameters:
        - in: query
          name: nom
          required: true
          schema: { type: string }
        - in: query
          name: prenom
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
        "400": { description: Requête invalide }
  /params/{nom}/{prenom}:
    get:
      summary: Exemple PathParams
      parameters:
        - in: path
          name: nom
          required: true
          schema: { type: string }
        - in: path
          name: prenom
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
  /login:
    post:
      summary: Authentification stateless
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        "200":
          description: Token généré
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        "401": { description: Identifiants invalides }
  /annonces:
    get:
      summary: Liste paginée des annonces
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 0, default: 0 }
        - in: query
          name: size
          schema: { type: integer, minimum: 1, default: 10 }
      responses:
        "200":
          description: Liste paginée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedAnnonceResponse'
        "401": { description: Non authentifié }
    post:
      summary: Créer une annonce
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnnonceDTO'
      responses:
        "201": { description: Créée }
        "400": { description: Validation }
        "401": { description: Non authentifié }
        "409": { description: Conflit métier }
  /annonces/{id}:
    get:
      summary: Détail d'une annonce
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200": { description: OK }
        "404": { description: Non trouvée }
    put:
      summary: Mettre à jour une annonce
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnnonceDTO'
      responses:
        "200": { description: Mise à jour }
        "403": { description: Interdit }
        "404": { description: Non trouvée }
        "409": { description: Conflit métier/version }
    delete:
      summary: Supprimer une annonce (après archivage)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "204": { description: Supprimée }
        "403": { description: Interdit }
        "404": { description: Non trouvée }
        "409": { description: Non archivée / conflit métier }
  /annonces/{id}/publish:
    post:
      summary: Publier une annonce
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200": { description: Publiée }
        "403": { description: Interdit }
        "404": { description: Non trouvée }
  /annonces/{id}/archive:
    post:
      summary: Archiver une annonce
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200": { description: Archivée }
        "403": { description: Interdit }
        "404": { description: Non trouvée }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-token
  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }
    LoginResponse:
      type: object
      properties:
        token: { type: string }
        userId: { type: integer, format: int64 }
        username: { type: string }
        message: { type: string }
    AnnonceDTO:
      type: object
      properties:
        id: { type: integer, format: int64 }
        titre: { type: string }
        description: { type: string }
        categorie: { type: string }
        prix: { type: number, format: double }
        statut: { type: string }
        auteurId: { type: integer, format: int64, readOnly: true }
        version: { type: integer, format: int64 }
    PagedAnnonceResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/AnnonceDTO'
        pageNumber: { type: integer }
        pageSize: { type: integer }
        totalElements: { type: integer, format: int64 }
        totalPages: { type: integer }
        hasNextPage: { type: boolean }