# src/main/java/fr/uit/univparis8/tpair/tpair1/dto/LoginRequest.java
package fr.uit.univparis8.tpair.tpair1.dto;


public class LoginRequest {
    public String username;
    public String password;

    public LoginRequest() {}

    public LoginRequest(String username, String password) {
        this.username = username;
        this.password = password;
    }
}